pyproject.toml
[tool.poetry]
name = "lee_links_backend"
version = "0.1.0"
description = ""
authors = ["Lee-Roy King <leeroywking@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
flask = "^2.0.2"
gunicorn = "^20.1.0"
python-dotenv = "^0.19.1"
spotipy = "^2.19.0"
flask_session = "^0.4.0"
flask_cors = "^3.0.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
